--- a/.eslintrc.js
+++ b/.eslintrc.js
@@ -96,7 +96,7 @@ module.exports = {
         'prettier',
     ],
     plugins: ['@typescript-eslint', 'jsdoc', 'you-dont-need-lodash-underscore', 'react-native-a11y', 'react', 'testing-library', 'eslint-plugin-react-compiler'],
-    ignorePatterns: ["lib/**"],
+    ignorePatterns: ['lib/**'],
     parser: '@typescript-eslint/parser',
     parserOptions: {
         project: path.resolve(__dirname, './tsconfig.json'),