--- a/.prettierignore
+++ b/.prettierignore
@@ -17,3 +17,5 @@ package-lock.json
 *.md
 # We need to modify the import here specifically, hence we disable prettier to get rid of the sorted imports
 src/libs/E2E/reactNativeLaunchingTest.ts
+# Temporary while we keep react-compiler in our repo
+lib/**