--- a/babel.config.js
+++ b/babel.config.js
@@ -3,7 +3,7 @@ require('dotenv').config();
 const IS_E2E_TESTING = process.env.E2E_TESTING === 'true';
 
 const ReactCompilerConfig = {
-    runtimeModule: "react-compiler-runtime",
+    runtimeModule: 'react-compiler-runtime',
 };
 const defaultPresets = ['@babel/preset-react', '@babel/preset-env', '@babel/preset-flow', '@babel/preset-typescript'];
 const defaultPlugins = [