--- a/babel.config.js
+++ b/babel.config.js
@@ -43,7 +43,6 @@ const webpack = {
 const metro = {
     presets: [require('@react-native/babel-preset')],
     plugins: [
-        ['babel-plugin-react-compiler', ReactCompilerConfig], // must run first!
         // This is needed due to a react-native bug: https://github.com/facebook/react-native/issues/29084#issuecomment-1030732709
         // It is included in metro-react-native-babel-preset but needs to be before plugin-proposal-class-properties or FlatList will break
         '@babel/plugin-transform-flow-strip-types',
@@ -145,6 +144,14 @@ module.exports = (api) => {
     // For `storybook` there won't be any config at all so we must give default argument of an empty object
     const runningIn = api.caller((args = {}) => args.name);
     console.debug('  - running in: ', runningIn);
+
+    // don't include react-compiler in jest, because otherwise tests will fail
+    if (runningIn !== 'babel-jest') {
+        // must run first!
+        metro.plugins.unshift(['babel-plugin-react-compiler', ReactCompilerConfig]);
+    }
+
+    console.log(metro);
 
     return ['metro', 'babel-jest'].includes(runningIn) ? metro : webpack;
 };