--- a/lib/react-compiler-runtime/package.json
+++ b/lib/react-compiler-runtime/package.json
@@ -1,0 +1,10 @@
+{
+  "name": "react-forget-runtime",
+  "version": "0.0.1",
+  "description": "Runtime for React Forget",
+  "license": "MIT",
+  "main": "index.js",
+  "dependencies": {
+    "react": "^18.2.0"
+  }
+}