… .github/actions/javascript/authorChecklist/index.js