@kirillzyusko 2024-06-04
@mountiny what are the next steps here? I remember you wanted to discuss it on Monday so it's a friendly reminder 👀 

My understanding is that it's pretty stable and we have a control of each file (so we always can disable unnecessary optimizations if needed).