@AndrewGable 2024-06-04
Should we run an AdHoc regression test to see any thing breaks? It seems like it might be a decent idea