@roryabraham 2024-06-04
Also, does this mean that we should start changing the way we write code such that we don't use `useMemo`, `useCallback`, etc... and just let the React Compiler handle memoization on its own?