@roryabraham 2024-06-04
I feel like the real test here would be to take some optimized but complex/less-readable code in an important component, remove all memoization, and then see how well the React Compiler does to automatically add memoization.