@kirillzyusko 2024-06-05
> There are no changes for the failing test on main, can you see if the npm run gh-actions-build command does produce a diff on this branch? I think we should commit those changes if so.

@AndrewGable I pulled lates main again, and it produced a different output again. I committed modified file, but may I ask you to shed some light on why it happens? I simply added new dependency to `package.json` - why it affected a compilation of GH action file? 🤔