@kirillzyusko 2024-06-05
> Should we run an AdHoc regression test to see any thing breaks? It seems like it might be a decent idea

@mountiny may I kindly ask you to re-trigger a new build? I think for a previous build I haven't updated a branch for a pretty long time, so I assume it may have old bugs. Now I rebased all my changes, so it should be in more up-to-date state (but if you think that we can test previous build, then I'm not against it).