@lanitochka17 2024-06-15
##  Setting - App crashes after click on "Settings > Share Code"

**Version Number:** 1.4.78-4
**Reproducible in staging?:** N
**Reproducible in production?:** N
**Issue reported by:** Applause-Internal team


## Action Performed:
1. Open Web or iOS App / 42287 Special Build
2. Login with any account
3. Go to Settings
4. Click / Tap on "Share Code" button

## Expected Result:
"Share Code" page is оpened

## Actual Result:
App crashes after click on "Settings > Share Code"

## Workaround:
Unknown


## Platforms:
<!---
Check off any platforms that are affected by this issue
--->
Which of our officially supported platforms is this issue occurring on?

- [ ] iOS: Native
- [x] MacOS: Chrome / Safari


## Screenshots/Videos

https://github.com/Expensify/App/assets/78819774/d28aa3a1-6b7d-4a2f-b88d-e2c7158ab598