@m-natarajan 2024-06-15
If you haven’t already, check out our [contributing guidelines](https://github.com/Expensify/ReactNativeChat/blob/main/contributingGuides/CONTRIBUTING.md) for onboarding and email contributors@expensify.com to request to join our Slack channel!
___
## Composer - Edit composer is not focused after canceling The delete confirmation
**Version Number:** 1.4.78-4
**Reproducible in staging?:** N
**Reproducible in production?:** N
**Issue reported by:** Applause-Internal team


## Action Performed:
1. Navigate to a conversation
2. Send a message
3. Click on edit from the action menu.
4. Clear all the texts in the edit composer
5. Click on the send button found on the edit composer
6. Notice that a delete confirmation window is shown. Click on Cancel
## Expected Result:
After canceling, the edit the composer is focused
## Actual Result:
After canceling, the main composer is focused
## Workaround:
Unknown


## Platforms:
<!---
Check off any platforms that are affected by this issue
--->
Which of our officially supported platforms is this issue occurring on?

- [ ] iOS: Native
- [x] MacOS: Chrome / Safari


## Screenshots/Videos

https://github.com/Expensify/App/assets/38435837/6c152ade-77e8-43e2-be7e-d878cafa146d



</details>

[View all open jobs on GitHub](https://github.com/Expensify/App/issues?q=is%3Aopen+is%3Aissue+label%3A%22Help+Wanted%22)