@kirillzyusko 2024-06-16
> Do you think you could sync with main

@mountiny done 🙌 

@lanitochka17 @m-natarajan I fixed reported issues as well 👀 (reported by @lanitochka17 indeed was caused by `react-compiler` usage, and reported issue by @m-natarajan was fixed by syncing with latest `main`).