@mountiny 2024-06-19
@kirillzyusko still around 30% of QA left