@kavimuru 2024-06-19
If you haven’t already, check out our [contributing guidelines](https://github.com/Expensify/ReactNativeChat/blob/main/contributingGuides/CONTRIBUTING.md) for onboarding and email contributors@expensify.com to request to join our Slack channel!
___
## Status Emoji menu is completely stretched
**Version Number:** 1.4.78-4
**Reproducible in staging?:** N
**Reproducible in production?:** N
**Issue reported by:** Applause-Internal team


## Action Performed:
1. Access https://42287.pr-testing.expensify.com/
2. Sign into a valid account
3. Go to Profile > Status > Emoji dropdown

## Expected Result:
User expects the emoji picker to fit the screen
## Actual Result:
The Emoji picked is completely stretched
## Workaround:
Unknown


## Platforms:
<!---
Check off any platforms that are affected by this issue
--->
Which of our officially supported platforms is this issue occurring on?

- [ ] iOS: Native
- [x] MacOS: Chrome / Safari


## Screenshots/Videos

https://github.com/Expensify/App/assets/43996225/6b95bbb6-a52f-4282-93aa-9453b953ac89



</details>

[View all open jobs on GitHub](https://github.com/Expensify/App/issues?q=is%3Aopen+is%3Aissue+label%3A%22Help+Wanted%22)