@kirillzyusko 2024-06-19
> still around 30% of QA left

@mountiny okay, can you ping me when QA finishes all checks so that I can go through the issues and squash all of them out?