@kirillzyusko 2024-06-19
But regarding newly reported issues:
- https://github.com/Expensify/App/pull/42287#issuecomment-2178796787 is not reproducible on a fresh build:

> <img width="1728" alt="image" src="https://github.com/Expensify/App/assets/22820318/adc71232-f9a0-43dd-8391-44d17bf257d9">

- https://github.com/Expensify/App/pull/42287#issuecomment-2178698317

> I think it's common problem that assets are not loading? I'm getting 404 error:
> <img width="792" alt="image" src="https://github.com/Expensify/App/assets/22820318/93f472a4-de61-462b-86ed-12351bb26dea">
> And I also tested other PR previews (for example this https://github.com/Expensify/App/pull/43993) and it has the same problem. So I tend to think that it's probably BE misconfiguration...